requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
markdown-pdf
python-docx